Reverse String
Reverse String
This Java program reverses a string using a recursive approach. The input string is given as an array of characters.

Working
The code consists of a single class ReverseString with a reverseString method. The approach employs recursion to reverse the string in-place.

Method: reverseString
Parameters:
s (char[]): Represents the input string as an array of characters.
Recursive Helper Function:
reverseHelper: Recursively swaps characters at the left and right ends of the string until the left pointer crosses the right pointer.
Base Case:
If the left pointer crosses the right pointer, the recursion stops.
Swap Operation:
Characters at the left and right positions are swapped using a temporary variable.
Recursion:
Recurs for the substring excluding the swapped characters until the entire string is reversed.
Example:
For input ["h","e","l","l","o"], the output will be ["o","l","l","e","h"].